Move get_lat_longs to meta_util
Avoids circular import that previously required portal parameter for
get_lat_longs
