+++ qiita_db/util.py
@@ -25,7 +25,6 @@ Methods
     check_required_columns
     convert_from_id
     convert_to_id
-    get_lat_longs
     get_environmental_packages
     purge_filepaths
     move_filepaths_to_upload_folder
@@ -967,42 +966,6 @@ def get_processed_params_tables():
                  ORDER BY table_name"""
         TRN.add(sql)
         return TRN.execute_fetchflatten()
-
-
-def get_lat_longs(portal):
-    """Retrieve the latitude and longitude of all the samples in the DB
-
-    Parameters
-    ----------
-    portal : qiita_db.portal.Portal object
-        The portal for which to get the lat_longs
-
-    Returns
-    -------
-    list of [float, float]
-        The latitude and longitude for each sample in the database
-    """
-    portal_table_ids = portal.get_studies()
-
-    with TRN:
-        sql = """SELECT DISTINCT table_name
-                 FROM information_schema.columns
-                 WHERE SUBSTR(table_name, 1, 7) = 'sample_'
-                    AND table_schema = 'qiita'
-                    AND column_name IN ('latitude', 'longitude');"""
-        TRN.add(sql)
-
-        sql = "SELECT latitude, longitude FROM qiita.{0}"
-        idx = TRN.index
-
-        portal_tables = TRN.execute_fetchflatten()
-        portal_tables = [x for x in TRN.execute_fetchflatten()
-                         if int(x[7:]) in portal_table_ids]
-
-        for table in portal_tables:
-            TRN.add(sql.format(table))
-
-        return list(chain.from_iterable(TRN.execute()[idx:]))
 
 
 def get_environmental_packages():
